1. Improved the code comments and made it more clear what's going on 2. Made sure that the tokenize result is properly cached 3. Added simple validation of the selectors that Sizzle considers valid (while they're not) 4. Added handling of invalid selectors in a stylesheet 5. Added unit-tests to check the points above
> Do we really need to change a behavior of tokenize when tolerant option is passed?

Yes, as long as we want to benefit from tokens caching mechanism. I'd love to move all the logic to the parser, though. But we have postpone it until we ditch Sizzle.

> Your point about caching makes sense, and I just realized that the current way prevents caching in general.

That one should be addressed now. I've verified that the caching works as intended.

> Ideally, we should be able to use tokens or compiled matcher function in Sizzle.select

The compiled matcher is created on Sizzle.compile call in the ExtendedSelector constructor, and it is cached against a selector created with toSelector call.